separate participant and auth
