+++ frontend/frontend-doc.ts
@@ -12,8 +12,9 @@ import { activateExamComponents, activateExamContent, setupCodeEditors } from ".
 import "./frontend-doc.css";
 import { assert } from '../src/core/util';
 import { AssignedExam } from '../src/core/assigned_exams';
-import { ExamCompletion } from '../src/verifiers/ExamCompletion';
+import { ExamCompletion } from './plugins/ExamCompletion';
 import { jwtDecode } from 'jwt-decode';
+import { Participant } from './plugins/Participant';
 
 
 function extractQuestionAnswers(question_elem: JQuery) : OpaqueQuestionAnswer {
@@ -469,14 +470,15 @@ async function startExam() {
   const manifest = parseExamManifest(exam_manifest_response.data);
     assert(isTransparentExamManifest(manifest));
     const assigned_exam = AssignedExam.createFromSubmission(exam, fillManifest(manifest, extractExamAnswers()));
+    const participant = new Participant(assigned_exam);
     
-    const completion = exam.completion && new ExamCompletion(assigned_exam, $("#examma-ray-exam-completion-status"));
+    const completion = exam.completion && new ExamCompletion(participant, $("#examma-ray-exam-completion-status"));
 
     (window as any).google_sign_in_callback = (response: any) => {
       const google_id_token = response.credential;
-      
+      assert(assigned_exam.exam.credentials_strategy); // otherwise callback should never have happened
       if (google_id_token && completion) {
-        completion.signIn(google_id_token);
+        participant.signIn(google_id_token, assigned_exam.exam.credentials_strategy.auth_endpoint);
       }
     }
 
@@ -507,7 +509,7 @@ async function startExam() {
         }
       })
 
-      completion?.verify();
+      completion?.refresh();
 
       first = false;
     };
