+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "examma-ray",
-  "version": "1.0.376",
+  "version": "1.0.377",
   "description": "## Setup",
   "main": "dist/core.js",
   "types": "dist/core.d.ts",
